feat(factions): DM sheet + edit toggle for FactionDetailView (#168)
Third item off the #168 audit list. `FactionDetailView` was a
monolithic form that owned both the editor fields and the linked
sub-sections (members, relations, locations, items, notes). Split it
into:

- `FactionEditor.vue` — core form fields (emblem upload, type +
  alignment comboboxes, visibility toggle, tags, name, Tiptap
  description) + Save/Cancel/Delete
- `FactionSheet.vue` — read-only DM view (emblem via FocalImage,
  name, type/alignment/tags blocks, Tiptap description via
  RichTextViewer) + Edit/Delete
- `FactionDetailView.vue` — thin route wrapper branching on
  isNew / `?edit=true`; the shared sub-sections (FactionMembersSection,
  FactionPartyMembersSection, FactionRelationsSection,
  FactionLocationsSection, FactionItemsSection, EntityNotesPanel)
  render once below the sheet/editor block since they're self-
  managing CRUD panels equally useful in both modes — no duplication.

Editor gains a **Cancel** button that strips `?edit=true` and
preserves other query params. Save / Delete navigation unchanged
(still `/factions`).
