+++ src/components/factions/FactionEditor.vue
@@ -1,0 +1,265 @@
+<template>
+  <div class="grid grid-cols-1 lg:grid-cols-[240px_1fr] gap-5">
+    <!-- Left: emblem + meta -->
+    <div class="flex flex-col gap-4">
+      <!-- Emblem -->
+      <div
+        class="relative aspect-square rounded-lg border border-border overflow-hidden bg-muted cursor-pointer group"
+        @click="fileInput?.click()"
+      >
+        <img
+          v-if="form.emblem_url"
+          :src="form.emblem_url"
+          alt="Faction emblem"
+          class="w-full h-full object-cover"
+        />
+        <div
+          v-else
+          class="w-full h-full flex flex-col items-center justify-center gap-2 text-muted-foreground"
+        >
+          <Shield class="h-10 w-10" />
+          <span class="font-fell text-sm italic">Upload emblem</span>
+        </div>
+        <div
+          class="absolute inset-0 bg-black/50 opacity-0 group-hover:opacity-100 transition-opacity flex items-center justify-center"
+        >
+          <span class="font-fell text-white text-sm italic">{{
+            form.emblem_url ? "Change" : "Upload"
+          }}</span>
+        </div>
+        <div
+          v-if="uploading"
+          class="absolute inset-0 bg-black/60 flex items-center justify-center"
+        >
+          <LoadingSpinner />
+        </div>
+      </div>
+      <input
+        ref="fileInput"
+        type="file"
+        accept="image/*"
+        class="hidden"
+        @change="onFileSelected"
+      />
+      <button
+        v-if="form.emblem_url"
+        type="button"
+        class="font-cinzel text-[10px] text-destructive hover:underline text-left"
+        @click.stop="form.emblem_url = ''"
+      >
+        Remove emblem
+      </button>
+
+      <!-- Type -->
+      <div class="space-y-1.5">
+        <label class="font-cinzel text-[10px] font-semibold tracking-wider text-muted-foreground uppercase">Type</label>
+        <EntityCombobox
+          v-model="factionTypeStr"
+          :options="FACTION_TYPE_OPTIONS"
+          placeholder="Select type…"
+        />
+      </div>
+
+      <!-- Alignment -->
+      <div class="space-y-1.5">
+        <label class="font-cinzel text-[10px] font-semibold tracking-wider text-muted-foreground uppercase">Alignment</label>
+        <EntityCombobox
+          v-model="alignmentStr"
+          :options="FACTION_ALIGNMENT_OPTIONS"
+          placeholder="Select alignment…"
+        />
+      </div>
+
+      <!-- Visibility -->
+      <div class="space-y-1.5">
+        <label class="font-cinzel text-[10px] font-semibold tracking-wider text-muted-foreground uppercase">Visible to Players</label>
+        <PlayerVisibilityToggle
+          :visible-to="form.player_visible_to"
+          @update:visible-to="form.player_visible_to = $event"
+        />
+      </div>
+
+      <!-- Tags -->
+      <div class="space-y-1.5">
+        <label class="font-cinzel text-[10px] font-semibold tracking-wider text-muted-foreground uppercase">Tags</label>
+        <TagInput v-model="tags" />
+      </div>
+    </div>
+
+    <!-- Right: name + description -->
+    <div class="flex flex-col gap-4">
+      <div class="space-y-1.5">
+        <label class="font-cinzel text-[10px] font-semibold tracking-wider text-muted-foreground uppercase">Name</label>
+        <input
+          v-model="form.name"
+          placeholder="Faction name…"
+          required
+          class="w-full bg-card border border-border rounded-md px-3 py-2 font-cinzel text-lg font-bold text-foreground placeholder:text-muted-foreground focus:outline-none focus:ring-1 focus:ring-ring"
+        />
+      </div>
+      <div class="flex-1 space-y-1.5">
+        <label class="font-cinzel text-[10px] font-semibold tracking-wider text-muted-foreground uppercase">Description & Notes</label>
+        <RichTextEditor
+          v-model="form.description"
+          placeholder="History, motives, known activities…"
+          min-height="320px"
+        />
+      </div>
+
+      <!-- Action row -->
+      <div class="flex items-center justify-end gap-2">
+        <button
+          v-if="!isNew"
+          type="button"
+          class="inline-flex items-center gap-1.5 rounded-md border border-destructive px-3 py-2 font-cinzel text-xs font-semibold text-destructive hover:bg-destructive/10 transition-colors"
+          @click="handleDelete"
+        >
+          <Trash2 class="h-3.5 w-3.5" />
+          Delete
+        </button>
+        <button
+          v-if="!isNew"
+          type="button"
+          class="inline-flex items-center gap-1.5 rounded-md border border-border px-3 py-2 font-cinzel text-xs font-semibold text-muted-foreground hover:text-foreground hover:border-muted-foreground/50 transition-colors"
+          @click="onCancel"
+        >
+          Cancel
+        </button>
+        <button
+          type="button"
+          :disabled="saving || !form.name.trim()"
+          class="inline-flex items-center gap-1.5 rounded-md bg-primary px-4 py-2 font-cinzel text-xs font-semibold text-primary-foreground tracking-wider hover:opacity-90 transition-opacity disabled:opacity-50"
+          @click="handleSave"
+        >
+          <Save class="h-3.5 w-3.5" />
+          {{ saving ? "Saving…" : isNew ? "Create" : "Save" }}
+        </button>
+      </div>
+    </div>
+  </div>
+</template>
+
+<script setup lang="ts">
+import { ref, computed, watch } from "vue";
+import { useRoute, useRouter } from "vue-router";
+import { Save, Shield, Trash2 } from "lucide-vue-next";
+import { useConfirm } from "@/composables/useConfirm";
+import { useImageUpload } from "@/composables/useImageUpload";
+import {
+  useCreateFaction,
+  useUpdateFaction,
+  useDeleteFaction,
+} from "@/composables/useFactions";
+import { FACTION_TYPES, FACTION_ALIGNMENTS, type Faction } from "@/types/faction.types";
+import LoadingSpinner from "@/components/common/LoadingSpinner.vue";
+import TagInput from "@/components/common/TagInput.vue";
+import RichTextEditor from "@/components/common/RichTextEditor.vue";
+import EntityCombobox from "@/components/common/EntityCombobox.vue";
+import PlayerVisibilityToggle from "@/components/common/PlayerVisibilityToggle.vue";
+
+const props = defineProps<{
+  faction: Faction | null;
+  isNew: boolean;
+}>();
+
+const route  = useRoute();
+const router = useRouter();
+const { confirm } = useConfirm();
+
+// Combobox option lists
+const FACTION_TYPE_OPTIONS = FACTION_TYPES.map((t) => ({ id: t, name: t }));
+const FACTION_ALIGNMENT_OPTIONS = FACTION_ALIGNMENTS.map((a) => ({ id: a, name: a }));
+
+const factionTypeStr = computed({
+  get: () => form.value.faction_type ?? "",
+  set: (v) => { form.value.faction_type = v || null; },
+});
+const alignmentStr = computed({
+  get: () => form.value.alignment ?? "",
+  set: (v) => { form.value.alignment = v || null; },
+});
+
+const createFaction = useCreateFaction();
+const updateFaction = useUpdateFaction();
+const deleteFaction = useDeleteFaction();
+const saving = ref(false);
+const uploading = ref(false);
+const fileInput = ref<HTMLInputElement | null>(null);
+const tags = ref<string[]>([]);
+
+const form = ref({
+  name: "",
+  faction_type: null as string | null,
+  description: null as string | null,
+  emblem_url: "" as string,
+  alignment: null as string | null,
+  player_visible_to: [] as string[],
+});
+
+watch(
+  () => props.faction,
+  (f) => {
+    if (!f) return;
+    form.value.name = f.name;
+    form.value.faction_type = f.faction_type;
+    form.value.description = f.description;
+    form.value.emblem_url = f.emblem_url ?? "";
+    form.value.alignment = f.alignment;
+    form.value.player_visible_to = f.player_visible_to ?? [];
+    tags.value = [...f.tags];
+  },
+  { immediate: true },
+);
+
+async function handleSave() {
+  if (!form.value.name.trim()) return;
+  saving.value = true;
+  try {
+    const payload = {
+      name: form.value.name.trim(),
+      faction_type: form.value.faction_type,
+      description: form.value.description,
+      emblem_url: form.value.emblem_url || null,
+      alignment: form.value.alignment,
+      player_visible_to: form.value.player_visible_to,
+      tags: tags.value,
+    };
+    if (props.isNew) {
+      await createFaction.mutateAsync(payload);
+    } else if (props.faction) {
+      await updateFaction.mutateAsync({ id: props.faction.id, update: payload });
+    }
+    router.push("/factions");
+  } finally {
+    saving.value = false;
+  }
+}
+
+async function handleDelete() {
+  if (!props.faction) return;
+  if (!(await confirm(`Delete "${props.faction.name}"? This cannot be undone.`))) return;
+  await deleteFaction.mutateAsync(props.faction.id);
+  router.push("/factions");
+}
+
+// Cancel strips ?edit=true, preserving any other query params.
+function onCancel() {
+  const { edit: _edit, ...rest } = route.query;
+  router.push({ query: rest });
+}
+
+const { upload: uploadEmblem } = useImageUpload("asset-images");
+
+async function onFileSelected(e: Event) {
+  const file = (e.target as HTMLInputElement).files?.[0];
+  if (!file) return;
+  (e.target as HTMLInputElement).value = "";
+  uploading.value = true;
+  try {
+    const url = await uploadEmblem(file);
+    if (url) form.value.emblem_url = url;
+  } finally {
+    uploading.value = false;
+  }
+}
+</script>
